requirements.txt
opencv-python-headless
imutils
numpy
pyyaml
Flask